Extract `MainNavbar` component
Would be responsible for picking between:

* regular navigation sidebar content
* launched app navigation sidebar content
